Refactoring the specs structure
